test(parser): anon binop and ambiguous sequence
